Check if user is student and screened on mentoring page
