+++ src/components/StudentCheck.tsx
@@ -1,0 +1,27 @@
+import { useContext, useEffect } from 'react';
+import { useHistory } from 'react-router-dom';
+import { ScreeningStatus } from '../types';
+import { ModalContext } from '../context/ModalContext';
+import { UserContext } from '../context/UserContext';
+
+const StudentCheck = () => {
+  const userContext = useContext(UserContext);
+  const modalContext = useContext(ModalContext);
+  const history = useHistory();
+
+  useEffect(() => {
+    if (userContext.user.type !== 'student') {
+      history.push('/');
+    }
+    if (
+      userContext.user.screeningStatus === ScreeningStatus.Unscreened ||
+      (userContext.user.isInstructor &&
+        userContext.user.instructorScreeningStatus ===
+          ScreeningStatus.Unscreened)
+    ) {
+      modalContext.setOpenedModal('accountNotScreened');
+    }
+  }, [userContext.user.screeningStatus]);
+};
+
+export default StudentCheck;
